requirements.txt
altgraph==0.17.4
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
CTkMessagebox==2.7
customtkinter==5.2.2
darkdetect==0.8.0
dateparser==1.2.1
docx==0.2.4
et_xmlfile==2.0.0
filelock==3.18.0
fsspec==2025.5.1
huggingface-hub==0.32.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
lxml==5.4.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pdf2image==1.17.0
pefile==2023.2.7
pillow==11.2.1
pycryptodome==3.23.0
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyMuPDF==1.26.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-pptx==1.0.2
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
six==1.17.0
soupsieve==2.7
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
XlsxWriter==3.2.3
